redo the boids enviornment
This commit is really big and does not make sense to split to small
commits as its a redo of the whole boids enviornment. Hence, everything
is commited at once.

This will serve as the starting point of the new boids env and from
here, it will be commited to gradualy in smaller commits.